This commit turns the logging module from a lame print(char) to an awesome printf style logging module. Now you can just call it like this:
logging_debug_high("user %s chose table %i\n", user, table_id);

and enjoy nice, easy to use debugging functions. logging takes care of if logging to screen / writing to a log file is on or off and whether the debugging meets the debug_level filter set in the config file too.

This really is the ideal logging module.